# build: reduce bundle size and enable code splitting 