--- a/src/webui/src/router.js
+++ b/src/webui/src/router.js
@@ -1,10 +1,12 @@
 import React from 'react';
 import {HashRouter as Router, Route, Switch} from 'react-router-dom';
+import {asyncComponent} from './utils/asyncComponent';
 
 import Header from './components/Header';
-import Home from './modules/home';
-import Detail from './modules/detail';
 import Footer from './components/Footer';
+
+const DetailPackage = asyncComponent(() => import('./modules/detail'));
+const HomePage = asyncComponent(() => import('./modules/home'));
 
 const RouterApp = () => {
   return (
@@ -13,9 +15,9 @@ const RouterApp = () => {
         <Header/>
         <div className="container">
           <Switch>
-            <Route exact path="/(search/:keyword)?" component={ Home } />
-            <Route exact path="/detail/@:scope/:package" component={Detail} />
-            <Route exact path="/detail/:package" component={Detail} />
+            <Route exact path="/(search/:keyword)?" component={ HomePage } />
+            <Route exact path="/detail/@:scope/:package" component={DetailPackage} />
+            <Route exact path="/detail/:package" component={DetailPackage} />
           </Switch>
         </div>
         <Footer/>