### Description

* fix some `lodash` dependencies where fetching the whole library (it saves raw ~100kb)
* enable dynamic import for routers and webpack produces chunks
* update webpack to `4.10.x`

⚠️  `element-react` still is using around 500kb and we use only couple of components, it should be smaller. There is still room for improvement there.
![screen shot 2018-06-03 at 6 42 50 pm](https://user-images.githubusercontent.com/558752/40888876-feaebb1a-675d-11e8-80ee-bc51071f53f6.png)
This bundle graph represents how it looks like after optimizations 


### Webpack output 

Before optimization

```
Hash: 859e1f6b0a14bba51c5f
Version: webpack 4.8.3
Time: 20016ms
Built at: 2018-06-03 18:37:15
                         Asset       Size  Chunks                    Chunk Names
                      logo.png   12.6 KiB          [emitted]         
style.2adec3d1fce0ae26f1fb.css    177 KiB       0  [emitted]         main
  main.859e1f6b0a14bba51c5f.js    928 KiB       0  [emitted]  [big]  main
                   favicon.ico   14.7 KiB          [emitted]         
                    index.html  618 bytes          [emitted]         
Entrypoint main [big] = style.2adec3d1fce0ae26f1fb.css main.859e1f6b0a14bba51c5f.js
 [14] (webpack)/buildin/module.js 519 bytes {0} [built]
[103] (webpack)/buildin/global.js 509 bytes {0} [built]
[105] ./src/webui/utils/api.js 2.5 KiB {0} [built]
[134] ./node_modules/react-router-dom/es/index.js + 30 modules 75.7 KiB {0} [built]
      |    31 modules
[138] ./src/webui/utils/url.js 1.09 KiB {0} [built]
[201] ./src/webui/utils/storage.js 1.11 KiB {0} [built]
[283] ./src/webui/src/styles/global.scss 39 bytes {0} [built]
[339] ./src/webui/src/modules/detail/index.jsx 7.89 KiB {0} [built]
[393] ./src/webui/src/modules/home/index.js 8.14 KiB {0} [built]
[397] ./src/lib/constants.js 692 bytes {0} [built]
[421] ./src/webui/src/router.js 2.22 KiB {0} [built]
[627] ./src/webui/src/app.js 2.43 KiB {0} [built]
[639] ./src/webui/utils/__setPublicPath__.js 149 bytes {0} [built]
[640] ./src/webui/src/index.js 1.39 KiB {0} [built]
[844] multi babel-polyfill whatwg-fetch ./src/webui/src/index.js 52 bytes {0} [built]
    + 830 hidden modules

WARNING in asset size limit: The following asset(s) exceed the recommended size limit (244 KiB).
This can impact web performance.
Assets: 
  main.859e1f6b0a14bba51c5f.js (928 KiB)

WARNING in entrypoint size limit: The following entrypoint(s) combined asset size exceeds the recommended limit (244 KiB). This can impact web performance.
Entrypoints:
  main (1.08 MiB)
      style.2adec3d1fce0ae26f1fb.css
      main.859e1f6b0a14bba51c5f.js


WARNING in webpack performance recommendations: 
You can limit the size of your bundles by using import() or require.ensure to lazy load some parts of your application.
For more info visit https://webpack.js.org/guides/code-splitting/
✨  Done in 26.34s.
```
After optimization

````
Hash: a4a5b717435dcd9a7e26
Version: webpack 4.10.2
Time: 20791ms
Built at: 2018-06-03 18:30:07
                           Asset       Size  Chunks                    Chunk Names
       3.a4a5b717435dcd9a7e26.js   15.2 KiB       3  [emitted]         
                        logo.png   12.6 KiB          [emitted]         
 vendors.a4a5b717435dcd9a7e26.js    727 KiB       0  [emitted]  [big]  vendors
       1.a4a5b717435dcd9a7e26.js   30.6 KiB       1  [emitted]         
2.style.7dc9de86de629d2c9353.css   8.69 KiB       2  [emitted]         
       2.a4a5b717435dcd9a7e26.js     30 KiB       2  [emitted]         
3.style.2e35f66ea7c26639c181.css   2.54 KiB       3  [emitted]         
0.style.b86c5878ee2935af83e9.css    159 KiB       0  [emitted]         vendors
4.style.0019dd8c1d27dc907ed8.css   3.19 KiB       4  [emitted]         
       4.a4a5b717435dcd9a7e26.js     26 KiB       4  [emitted]         
manifest.a4a5b717435dcd9a7e26.js   3.02 KiB       5  [emitted]         manifest
6.style.2002218866f686476641.css   3.83 KiB       6  [emitted]         main
    main.a4a5b717435dcd9a7e26.js   55.4 KiB       6  [emitted]         main
                     favicon.ico   14.7 KiB          [emitted]         
                      index.html  900 bytes          [emitted]         
 [40] (webpack)/buildin/module.js 497 bytes {0} [built]
[123] (webpack)/buildin/global.js 489 bytes {0} [built]
[179] ./node_modules/react-router-dom/es/index.js + 30 modules 75.7 KiB {0} [built]
      |    31 modules
[181] ./src/webui/utils/storage.js 1.11 KiB {6} [built]
[258] ./src/webui/utils/api.js 2.53 KiB {6} [built]
[261] ./src/webui/utils/url.js 1.09 KiB {6} [built]
[269] ./src/webui/src/styles/global.scss 39 bytes {6} [built]
[306] ./src/webui/src/utils/asyncComponent.js 3.2 KiB {6} [built]
[308] ./src/webui/src/router.js 2.54 KiB {6} [built]
[514] ./src/webui/src/app.js 2.47 KiB {6} [built]
[526] ./src/webui/utils/__setPublicPath__.js 149 bytes {6} [built]
[527] ./src/webui/src/index.js 1.39 KiB {6} [built]
[731] multi babel-polyfill whatwg-fetch ./src/webui/src/index.js 52 bytes {6} [built]
[732] ./src/webui/src/modules/home/index.js 8.18 KiB {3} [built]
[733] ./src/webui/src/modules/detail/index.jsx 7.93 KiB {4} [built]
    + 929 hidden modules

WARNING in asset size limit: The following asset(s) exceed the recommended size limit (244 KiB).
This can impact web performance.
Assets: 
  vendors.a4a5b717435dcd9a7e26.js (727 KiB)

WARNING in entrypoint size limit: The following entrypoint(s) combined asset size exceeds the recommended limit (244 KiB). This can impact web performance.
Entrypoints:
  main (947 KiB)
      manifest.a4a5b717435dcd9a7e26.js
      0.style.b86c5878ee2935af83e9.css
      vendors.a4a5b717435dcd9a7e26.js
      6.style.2002218866f686476641.css
      main.a4a5b717435dcd9a7e26.js

✨  Done in 26.60s.
```